Converted to multi-module project
