--- src/test/java/com/example/websocket_chat/WebsocketChatApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.websocket_chat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WebsocketChatApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
